requirements.txt
azure-identity==1.15.0
azure-ai-ml==1.12.1
scikit-learn==1.3.2
pandas==2.1.4
pyarrow==14.0.2
python-dotenv==1.0.0
mldesigner==0.1.0b17 
azureml-mlflow==1.54.0.post1
matplotlib==3.8.2